extract storage hooks
